requirements.txt
django==1.11.13
django-filter==1.1.0
djangorestframework==3.8.2
requests==2.17.3
celery==4.1.1
mysqlclient==1.3.12
django_cors_headers==2.1.0
coverage==4.4.1
django_nose==1.4.4
django-mysql==2.3.0
gunicorn==19.8.1
